+++ compiler-rt/lib/rtsan/rtsan_diagnostics.cpp
@@ -91,5 +91,10 @@ void __rtsan::PrintDiagnostics(const DiagnosticsInfo &info) {
 void __rtsan::PrintErrorSummary(const DiagnosticsInfo &info,
                                 const BufferedStackTrace &stack) {
   ScopedErrorReportLock::CheckLocked();
-  ReportErrorSummary(GetErrorTypeStr(info), &stack);
+
+  // If we don't explicitly specify the tool name here, UBSan can
+  // report our error summaries with "UndefinedBehaviorSanitizer" prefix
+  // depending on initialization.
+  const char *realtimeSanitizerName = "RealtimeSanitizer";
+  ReportErrorSummary(GetErrorTypeStr(info), &stack, realtimeSanitizerName);
 }