+++ llvm/docs/ReleaseNotes.md
@@ -307,6 +307,7 @@ Changes to Sanitizers
 ---------------------
 
 * Add a random delay into ThreadSanitizer to help find rare thread interleavings.
+* Support running RealtimeSanitizer with UndefinedBehaviorSanitizer
 
 Other Changes
 -------------